pyproject.toml
[project]
name = "graphiti-knowledge-graph"
version = "0.1.0"
description = "Dual-scope knowledge graph system with global preferences and project-specific memory"
requires-python = ">=3.12"
dependencies = [
    "detect-secrets>=1.5.0",
    "GitPython>=3.1.0",
    "graphiti-core[kuzu]==0.26.3",
    "httpx>=0.28.0",
    "kuzu==0.11.3",
    "ollama==0.6.1",
    "persist-queue==1.1.0",
    "structlog>=25.5.0",
    "tenacity==9.1.2",
    "typer[all]>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
graphiti = "src.cli:cli_entry"
gk = "src.cli:cli_entry"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
